Move, slightly refactor, `handleEditGeneralGovernance`
