+++ src/pages/dao/settings/general/SafeGeneralSettingsPage.tsx
@@ -1,11 +1,8 @@
 import { Box, Flex, Show, Text } from '@chakra-ui/react';
-import { abis } from '@fractal-framework/fractal-contracts';
 import { useFormikContext } from 'formik';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
-import { encodeAbiParameters, encodeFunctionData, parseAbiParameters, zeroAddress } from 'viem';
-import { ZodiacModuleProxyFactoryAbi } from '../../../../assets/abi/ZodiacModuleProxyFactoryAbi';
+import { zeroAddress } from 'viem';
 import { GaslessVotingToggleDAOSettings } from '../../../../components/GaslessVoting/GaslessVotingToggle';
 import { SettingsContentBox } from '../../../../components/SafeSettings/SettingsContentBox';
 import { InputComponent } from '../../../../components/ui/forms/InputComponent';
@@ -17,20 +14,12 @@ import {
 import NestedPageHeader from '../../../../components/ui/page/Header/NestedPageHeader';
 import Divider from '../../../../components/ui/utils/Divider';
 import { DAO_ROUTES } from '../../../../constants/routes';
-import { usePaymasterDepositInfo } from '../../../../hooks/DAO/accountAbstraction/usePaymasterDepositInfo';
-import useSubmitProposal from '../../../../hooks/DAO/proposal/useSubmitProposal';
 import { useCurrentDAOKey } from '../../../../hooks/DAO/useCurrentDAOKey';
 import { useCanUserCreateProposal } from '../../../../hooks/utils/useCanUserSubmitProposal';
 import { createAccountSubstring } from '../../../../hooks/utils/useGetAccountName';
-import { useInstallVersionedVotingStrategy } from '../../../../hooks/utils/useInstallVersionedVotingStrategy';
 import { useStore } from '../../../../providers/App/AppProvider';
 import { useNetworkConfigStore } from '../../../../providers/NetworkConfig/useNetworkConfigStore';
-import { GovernanceType, ProposalExecuteData } from '../../../../types';
-import {
-  getPaymasterAddress,
-  getPaymasterSaltNonce,
-  getVoteSelectorAndValidator,
-} from '../../../../utils/gaslessVoting';
+import { GovernanceType } from '../../../../types';
 
 export function SafeGeneralSettingsPage() {
   const { t } = useTranslation('settings');
@@ -43,9 +32,8 @@ export function SafeGeneralSettingsPage() {
 
   const { daoKey } = useCurrentDAOKey();
   const {
-    governanceContracts: { strategies },
     governance: { type: votingStrategyType },
-    node: { subgraphInfo, safe, gaslessVotingEnabled, paymasterAddress },
+    node: { subgraphInfo, safe, gaslessVotingEnabled },
   } = useStore({ daoKey });
 
   const [existingIsGaslessVotingEnabledToggled, setExistingIsGaslessVotingEnabledToggled] =
@@ -55,19 +43,9 @@ export function SafeGeneralSettingsPage() {
     setExistingIsGaslessVotingEnabledToggled(gaslessVotingEnabled);
   }, [gaslessVotingEnabled]);
 
-  const navigate = useNavigate();
-
-  const { submitProposal } = useSubmitProposal();
   const { canUserCreateProposal } = useCanUserCreateProposal();
-  const {
-    addressPrefix,
-    chain: { id: chainId },
-    contracts: { keyValuePairs, accountAbstraction, paymaster, zodiacModuleProxyFactory },
-    bundlerMinimumStake,
-  } = useNetworkConfigStore();
-  const { depositInfo } = usePaymasterDepositInfo();
+  const { addressPrefix, bundlerMinimumStake } = useNetworkConfigStore();
   const accountAbstractionSupported = bundlerMinimumStake !== undefined;
-  const stakingRequired = accountAbstractionSupported && bundlerMinimumStake > 0n;
 
   const isMultisigGovernance = votingStrategyType === GovernanceType.MULTISIG;
   const gaslessVotingSupported = !isMultisigGovernance && accountAbstractionSupported;
@@ -88,10 +66,6 @@ export function SafeGeneralSettingsPage() {
     }
   }, [subgraphInfo?.daoName, subgraphInfo?.daoSnapshotENS, safeAddress]);
 
-  const nameChanged = !!existingDaoName && existingDaoName !== subgraphInfo?.daoName;
-  const snapshotChanged = existingSnapshotENS !== subgraphInfo?.daoSnapshotENS;
-  const gaslessVotingChanged = existingIsGaslessVotingEnabledToggled !== gaslessVotingEnabled;
-
   useEffect(() => {
     if (
       !formValues.general?.name &&
@@ -101,192 +75,6 @@ export function SafeGeneralSettingsPage() {
       setFieldValue('general', undefined);
     }
   }, [setFieldValue, formValues.general]);
-
-  const { buildInstallVersionedVotingStrategies } = useInstallVersionedVotingStrategy();
-
-  // @todo: Move logic to parent formik submit handler
-  const handleEditGeneralGovernance = async () => {
-    const changeTitles = [];
-    const keyArgs = [];
-    const valueArgs = [];
-
-    if (nameChanged) {
-      changeTitles.push(t('updatesSafeName', { ns: 'proposalMetadata' }));
-      keyArgs.push('daoName');
-      valueArgs.push(existingDaoName);
-    }
-
-    if (snapshotChanged) {
-      changeTitles.push(t('updateSnapshotSpace', { ns: 'proposalMetadata' }));
-      keyArgs.push('snapshotENS');
-      valueArgs.push(existingSnapshotENS);
-    }
-
-    if (gaslessVotingChanged) {
-      keyArgs.push('gaslessVotingEnabled');
-      if (existingIsGaslessVotingEnabledToggled) {
-        changeTitles.push(t('enableGaslessVoting', { ns: 'proposalMetadata' }));
-        valueArgs.push('true');
-      } else {
-        changeTitles.push(t('disableGaslessVoting', { ns: 'proposalMetadata' }));
-        valueArgs.push('false');
-      }
-    }
-
-    const title = changeTitles.join(`; `);
-
-    const targets = [keyValuePairs];
-    const calldatas = [
-      encodeFunctionData({
-        abi: abis.KeyValuePairs,
-        functionName: 'updateValues',
-        args: [keyArgs, valueArgs],
-      }),
-    ];
-    const values = [0n];
-
-    if (gaslessVotingChanged && existingIsGaslessVotingEnabledToggled) {
-      if (!safeAddress) {
-        throw new Error('Safe address is not set');
-      }
-
-      if (!accountAbstraction) {
-        throw new Error('Account Abstraction addresses are not set');
-      }
-
-      if (paymasterAddress === null) {
-        // Paymaster does not exist, deploy a new one
-        const paymasterInitData = encodeFunctionData({
-          abi: abis.DecentPaymasterV1,
-          functionName: 'initialize',
-          args: [
-            encodeAbiParameters(parseAbiParameters(['address', 'address', 'address']), [
-              safeAddress,
-              accountAbstraction.entryPointv07,
-              accountAbstraction.lightAccountFactory,
-            ]),
-          ],
-        });
-
-        targets.push(zodiacModuleProxyFactory);
-        calldatas.push(
-          encodeFunctionData({
-            abi: ZodiacModuleProxyFactoryAbi,
-            functionName: 'deployModule',
-            args: [
-              paymaster.decentPaymasterV1MasterCopy,
-              paymasterInitData,
-              getPaymasterSaltNonce(safeAddress, chainId),
-            ],
-          }),
-        );
-        values.push(0n);
-      }
-
-      // Include txs to disable any old voting strategies and enable the new ones.
-      const { installVersionedStrategyTxDatas, newStrategies } =
-        await buildInstallVersionedVotingStrategies();
-
-      targets.push(...installVersionedStrategyTxDatas.map(tx => tx.targetAddress));
-      calldatas.push(...installVersionedStrategyTxDatas.map(tx => tx.calldata));
-      values.push(...installVersionedStrategyTxDatas.map(() => 0n));
-
-      const predictedPaymasterAddress = getPaymasterAddress({
-        safeAddress,
-        zodiacModuleProxyFactory,
-        paymasterMastercopy: paymaster.decentPaymasterV1MasterCopy,
-        entryPoint: accountAbstraction.entryPointv07,
-        lightAccountFactory: accountAbstraction.lightAccountFactory,
-        chainId,
-      });
-
-      // Add stake for Paymaster if not enough
-      if (stakingRequired) {
-        const stakedAmount = depositInfo?.stake || 0n;
-
-        if (paymasterAddress === null || stakedAmount < bundlerMinimumStake) {
-          const delta = bundlerMinimumStake - stakedAmount;
-
-          targets.push(predictedPaymasterAddress);
-          calldatas.push(
-            encodeFunctionData({
-              abi: abis.DecentPaymasterV1,
-              functionName: 'addStake',
-              // one day in seconds, defined on https://github.com/alchemyplatform/rundler/blob/c17fd3dbc24d2af93fd68310031d445d5440794f/crates/sim/src/simulation/mod.rs#L170
-              args: [86400],
-            }),
-          );
-          values.push(delta);
-        }
-      }
-
-      newStrategies.forEach(strategy => {
-        // Whitelist the new strategy's `vote` function call on the Paymaster
-        // // // // // // // // // // // // // // // // // // // // // // //
-        const { voteSelector, voteValidator } = getVoteSelectorAndValidator(
-          strategy.type,
-          paymaster,
-        );
-
-        targets.push(predictedPaymasterAddress);
-        calldatas.push(
-          encodeFunctionData({
-            abi: abis.DecentPaymasterV1,
-            functionName: 'setFunctionValidator',
-            args: [strategy.address, voteSelector, voteValidator],
-          }),
-        );
-        values.push(0n);
-      });
-
-      // Also whitelist existing versioned strategies that have not already been whitelisted.
-      // This will be the case for DAOs deployed after this feature, but did not enable gasless voting on creation.
-      if (paymasterAddress === null) {
-        strategies
-          .filter(strategy => strategy.version !== undefined)
-          .forEach(strategy => {
-            const { voteSelector, voteValidator } = getVoteSelectorAndValidator(
-              strategy.type,
-              paymaster,
-            );
-
-            targets.push(predictedPaymasterAddress);
-            calldatas.push(
-              encodeFunctionData({
-                abi: abis.DecentPaymasterV1,
-                functionName: 'setFunctionValidator',
-                args: [strategy.address, voteSelector, voteValidator],
-              }),
-            );
-            values.push(0n);
-          });
-      }
-    }
-
-    const proposalData: ProposalExecuteData = {
-      metaData: {
-        title,
-        description: '',
-        documentationUrl: '',
-      },
-      targets,
-      values,
-      calldatas,
-    };
-    submitProposal({
-      safeAddress: safe?.address,
-      proposalData,
-      nonce: safe?.nextNonce,
-      pendingToastMessage: t('proposalCreatePendingToastMessage', { ns: 'proposal' }),
-      successToastMessage: t('proposalCreateSuccessToastMessage', { ns: 'proposal' }),
-      failedToastMessage: t('proposalCreateFailureToastMessage', { ns: 'proposal' }),
-      successCallback: () => {
-        if (safeAddress) {
-          navigate(DAO_ROUTES.dao.relative(addressPrefix, safeAddress));
-        }
-      },
-    });
-  };
 
   return (
     <>
